tests/wpt/tests/client-hints/accept-ch/meta/resource-in-markup-delegate-ch.https.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<meta http-equiv="Delegate-CH" content="sec-ch-dpr">

<script src="../../resources/script-set-dpr-header.py"></script>

<script>
test(() => {
  assert_greater_than(dprHeader.length, 0,
                      "sec-ch-dpr header should have been received");
}, "DPR is received in page with Accept-CH meta tag");
</script>
